autoHttp on changing temp&humi, check description
When temp change 1 °C or humidity 5% up/down, an HTTP request is automatically sent to the database. this is checked every 30seconds. Also a lot of changing rtos structure in this commit. Changed the reading sensor from RTOS task to a general function call. It has its pro's and cons that I just forgot... if you want to know, tell me and I will think about what it was. Then also some cleaning up code